Options {DOCKER_ROOT_VOLUME, DOCKER_ROOT_VOLUME_SIZE} for setting up separate logical volume for docker root.
At present, docker root, either default (/var/lib/docker) or
custom root location passed using -g or --graph option is stored
under the root (/) partition. This means all the docker related
content like images/containers/volumes are stored under `/` partition.
With this setup docker has the potential to exhaust the root partition.
This is not a problem for devicemapper if the user sets up a thin pool
(backed up by a volume group) for docker storage needs, since all the
images/containers would be stored on devices coming out of this pool
and not root `/` partition. For other graphdrivers e.g overlay this can
be a problem.

This patch allows a user to pass these two options, which would carve
out a separate logical volume (lv) from the volume group (vg) and
bind mount it on the docker root directory. With this setup in place,
docker content (images/containers/volumes) would come out of this lv
(vg) rather than exhausting the root partition.

Signed-off-by: Shishir Mahajan <shishir.mahajan@redhat.com>