Use wipefs -f to force remove signatures from the partition.
After partition creation, we wipe any signatures on the partition using wipefs.
For some reason, for loop devices it was detecting nested partition table signature
and wipefs will skip it by default and instead ask user whether to clean it or not.

Use option -f instead to not ask user the question and delete it always.

Signed-off-by: Shishir Mahajan <shishir.mahajan@redhat.com>